Elapsed Times, 'Enter' Event Button
Added elapsed times for start and end of songs to be used with computation in R. Added an event log entry for enter presses.

Big limitations found. The XML file from stepmania only updates when there is a high score on a song, so if you play a song and do not beat the previous score it will not be added to our JSON. The current work around is to delete the XML file before playing.